Handle output_directory and change_directory properly
If option change_directory is used output_directory should be relative
to the new pwd (in current implementation that would be `.` always) or
use absolute path. In addition if option change_directory is used flag
`-output-directory` doesn't need to be set on the tex compilation
command.

This implements this behaviour by setting `options.output_directory` to
`nil` (so that `-output-directory` will not be set later) after
generating the init lua script.