Change LUAINPUTS env var to find lua files on dir change
When changing directory, we already modify `TEXINPUTS` so that `.tex`
files are found from the original directory.
Since this doesn't work for `.lua` files we have to adjust another
environment variable as well, `LUAINPUTS`.
This is what this PR is for.

For Reference on the environment variables see `kpsewhich --help-formats`