Add option for sagetex
`sage` aims to be an alternative to commercial programs such as Maple,
Matlab, and Mathematica. `sagetex` is a package to include sage
commands/expressions in TeX documents and doing calculations/plotting
thereby.
Using `sagetex` requires run `tex -> sage -> tex`. As this workflow is
pretty similar to the one of biber/makeglossaries, we can implement it
quite the same way. So this just checks if `sage` should be run (file
with ext = `sage` found in `filelist` has changed) and if so, run
`sage`.